kula

+++ Assets/_Scripts/RobotMovement.cs
@@ -255,7 +255,7 @@ public class RobotMovement : MonoBehaviour
     {
         input += i;
     }
-    public void SetSpeedlvl(int leftSpeed, int rightSpeed)
+    public void SetSpeedlvl(float leftSpeed, float rightSpeed)
     {
         leftTrack.acceleration = leftTrack.acceleration*leftSpeed;
         rightTrack.acceleration = rightTrack.acceleration*rightSpeed;
